adds implementation code for employeeList
